+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * docs/guides/*: Added some basic guides on how to manage different resources with dependencies
 * docs/resources/subaccount.md: Add group_ids to example
 * provider/resources/index.go: Invoke a pause, validate state, and ensure all indexes are deleted on tear down
+* client/models.go: Extend out `Transforms` type to contain `queries` for JQ Transforms
+* docs/resources/view.md: Add an example JQ Transform
 
 ### Bug Fixes:
 * provider/resources/objectgroup.go: Fixed updates in place for `options` on apply